@@ -5,15 +5,10 @@ import com.example.creditSimulator.dto.LoanResponse
 import com.example.creditSimulator.entity.Client
 import com.example.creditSimulator.entity.InterestRate
 import com.example.creditSimulator.entity.Loan
-import org.springframework.boot.autoconfigure.integration.IntegrationProperties
-
 import org.springframework.stereotype.Service
-import java.math.BigDecimal
-
 
 @Service
 class LoanService {
-
     fun calculateLoan(loanRequest: LoanRequest): LoanResponse {
         val client = Client(birthDate = loanRequest.clientBirthDate)
 
@@ -21,11 +16,13 @@ class LoanService {
 
         loanInterestRate.rate
 
-        val loan = Loan(interestRate = loanInterestRate.rate,
-                        loanAmount = loanRequest.loanRequestAmount,
-                        paymentMonths = loanRequest.loanTime)
+        val loan =
+            Loan(
+                interestRate = loanInterestRate.rate,
+                loanAmount = loanRequest.loanRequestAmount,
+                paymentMonths = loanRequest.loanTime,
+            )
 
         return loan.calculateCredit()
-
     }
 }
